Introduce MockAutoloader
The MockAutoloader can be used to "overwrite" existing Magento
classes and replace them with local copies. This is needed as
not all interfaces expose all methods that can be called. Until
this is fixed in Magento the mechanism can make sure that
PHPStan knows about those methods and does not complain.