Remove autoloader magic in registration.php
Unfortunately made it impossible to overload the autoloader and
magically set the custom autoloaders needed for the extension to
run. Thus I removed the registration.php file handled the magic.

This means regardless of using the .phar archive or phpstan-src
dependency, you have to manually configure the autoload.php
work-a-round in the project's phpstan configuration.