Add Stubs for Magento's loosey-string behavior
In many cases, Magento typedefs a `string` but truly accepts anything that can be cast to a string. This can be annoying for things like Escaper and Session Messages, which many are inclined to pass Phrase objects into.

Additionally, Escaper has return types of `string|array`, passing an array if the input is an array. However, Magento's documentation blocks don't specify that's when it passes an array - so phpstan thinks that the result of Escaper can be either at any time.  In that way, you cannot echo out the result of an `escapeHtml` in a PHTML file on max level without getting some errors.